+++ python/lsst/ts/m2/translator.py
@@ -46,8 +46,6 @@ class Translator:
         match message_reformat["id"]:
             case "tangentForce":
                 message_reformat = self._handle_tangent_force(message_reformat)
-            case "commandableByDDS":
-                message_reformat = self._handle_commandable_by_dds(message_reformat)
             case "config":
                 message_reformat = self._handle_config(message_reformat)
             case "configurationFiles":
@@ -81,28 +79,6 @@ class Translator:
         """
 
         message["lutTemperature"] = [0.0] * 6
-
-        return message
-
-    def _handle_commandable_by_dds(self, message: dict) -> dict:
-        """Handle the message of commandable by data distribution system (DDS).
-
-        TODO: Remove this function after the DM-37422 is fixed. At this moment,
-        we need to connect the CSC to the ports of GUI to establish the
-        TCP/IP connection.
-
-        Parameters
-        ----------
-        message : `dict`
-            Message from the component.
-
-        Returns
-        -------
-        message : `dict`
-            Reformated message.
-        """
-
-        message["state"] = True
 
         return message